[Refactor] fetchFetchStatsをUtilsに移動
