+++ src/js/memoTrainingUtils.js
@@ -3,6 +3,7 @@ const constant = require('./constant');
 const numbering3 = require('./numbering3');
 const utils = require('./utils');
 const _ = require('lodash');
+const moment = require('moment-timezone');
 const shuffle = require('shuffle-array');
 const rp = require('request-promise');
 
@@ -161,6 +162,26 @@ export const fetchScore = (userName, event, mode) => {
             userName,
             event,
             mode,
+            token: localStorage.token,
+        },
+    };
+
+    return rp(options);
+};
+
+export const requestFetchStats = (userName, event, startDate, endDate) => {
+    const options = {
+        url: `${config.apiRoot}/getMemoLogStats`,
+        method: 'POST',
+        headers: {
+            'Content-Type': 'application/json',
+        },
+        json: true,
+        form: {
+            userName,
+            event,
+            startDate: moment(startDate, 'YYYY/MM/DD').toISOString(),
+            endDate: moment(endDate, 'YYYY/MM/DD').hour(23).minute(59).second(59).toISOString(),
             token: localStorage.token,
         },
     };
